requirements.txt
# Core RAG dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# Web crawler
crawl4ai>=0.2.0

# Quantum computing
pennylane>=0.33.0
qiskit>=0.45.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# NLP & Evaluation
nltk>=3.8.0

# API
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
